Move "Input" class into own file
