+++ src/midiplay.cc
@@ -19,6 +19,8 @@
 
 #include <assert.h>
 
+#include "../include/adlinput.h"
+
 #define SUPPORT_VIDEO_OUTPUT
 #define SUPPORT_PUZZLE_GAME
 
@@ -232,64 +234,7 @@ static const char PercussionMap[256] =
 "\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0"
 "\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0";
 
-class Input
-{
-#ifdef __WIN32__
-    void* inhandle;
-#endif
-#if (!defined(__WIN32__) || defined(__CYGWIN__)) && !defined(__DJGPP__)
-    struct termio back;
-#endif
-public:
-    Input()
-    {
-#ifdef __WIN32__
-        inhandle = GetStdHandle(STD_INPUT_HANDLE);
-#endif
-#if (!defined(__WIN32__) || defined(__CYGWIN__)) && !defined(__DJGPP__)
-        ioctl(0, TCGETA, &back);
-        struct termio term = back;
-        term.c_lflag &= ~(ICANON|ECHO);
-        term.c_cc[VMIN] = 0; // 0=no block, 1=do block
-        if(ioctl(0, TCSETA, &term) < 0)
-            fcntl(0, F_SETFL, fcntl(0, F_GETFL) | O_NONBLOCK);
-#endif
-    }
-    ~Input()
-    {
-#if (!defined(__WIN32__) || defined(__CYGWIN__)) && !defined(__DJGPP__)
-        if(ioctl(0, TCSETA, &back) < 0)
-            fcntl(0, F_SETFL, fcntl(0, F_GETFL) &~ O_NONBLOCK);
-#endif
-    }
-
-    char PeekInput()
-    {
-#ifdef __DJGPP__
-        if(kbhit()) { int c = getch(); return c ? c : getch(); }
-#endif
-#ifdef __WIN32__
-        DWORD nread=0;
-        INPUT_RECORD inbuf[1];
-        while(PeekConsoleInput(inhandle,inbuf,sizeof(inbuf)/sizeof(*inbuf),&nread)&&nread)
-        {
-            ReadConsoleInput(inhandle,inbuf,sizeof(inbuf)/sizeof(*inbuf),&nread);
-            if(inbuf[0].EventType==KEY_EVENT
-            && inbuf[0].Event.KeyEvent.bKeyDown)
-            {
-                char c = inbuf[0].Event.KeyEvent.uChar.AsciiChar;
-                unsigned s = inbuf[0].Event.KeyEvent.wVirtualScanCode;
-                if(c == 0) c = s;
-                return c;
-        }   }
-#endif
-#if (!defined(__WIN32__) || defined(__CYGWIN__)) && !defined(__DJGPP__)
-        char c = 0;
-        if(read(0, &c, 1) == 1) return c;
-#endif
-        return '\0';
-    }
-} Input;
+ADL_Input Input;
 
 #ifdef SUPPORT_PUZZLE_GAME
 #include "puzzlegame.inc"
